makepkg.conf: Set MAKEFLAGS dynamically
MAKEFLAGS="-j -l $(nproc)"

Since makepkg.conf is sourced by a bash script, I should be able to set
it dynamically. Also, leave -j empty to leave it unbounded but instead
use -l $(nproc) to bound by actual load.

This might all not be so relevant because not as many packages are using
make(1) these days.