.config/sway: Remove bg and bar references to /home/matt
This still isn't great, swaybg_command is apparently relative to the
config, and swaybar_command to the original working dir or home (looking
at chdir and execve syscalls through strace.

I filed https://github.com/swaywm/sway/issues/5724 to see if it would be
acceptable to support richer commands for these that don't need a
wrapper. For now, this is a little better since it means my dotfiles can
be used for other accounts more readily.