waybar: Add padding hack for ultrawide
This add some padding to either end of an ultrawide so that the max
usable width of the bar is that of a single 2k monitor, centered. In a
multi-screen setup like when I have the laptop open while connected to
the ultrawide, the padding only affects the ultrawide. It does this by
relying on the module being skipped when there isn't anything in the
format string, othwise it prints a zero-width space character.

I like waybar but man is the config clunky. Some built in jsonnet
support with an optional script to collect environmental data would go a
loooooong way.