firefox: More firejail enforcement, set as default (properly)
Add a `firefox` wrapper to $PATH to help enforce firejail.

Take in and override `firefox.desktop` in order to use this from xdg
things... Unfortunately the upstream firefox.desktop uses
`/usr/lib/firefox/firefox` as it's `Exec=` line.

Then start tracking `mimeapps.list` and `firefox.desktop`. With this:

```
$ xdg-settings check default-web-browser firefox.desktop
yes
```

this checks all of the default for http, https, and html.

Now, unfortunately it looks like firefox's check that it's the default
parses the .desktop Exec line looking for the /usr/lib/firefox/firefox
path so we still need to tell firefox to stop checking that it's the
default.... Even though it is.